Cambiado el tipo a int para cumplir con los argumentos de analogWrite

+++ lib/AnalogOutput/AnalogOutput.cpp
@@ -4,6 +4,6 @@ AnalogOutput::AnalogOutput(uint8_t pin)
     : Pin{ pin, OUTPUT } {
 }
 
-void AnalogOutput::write(uint8_t value) {
+void AnalogOutput::write(uint16_t value) {
     analogWrite(this->getPin(), value);
 }
